Make DGML viewer required Fix Scale bug with varchar
fixes #527
fixes #525

+++ src/GUI/RevEng.Core/Procedures/Model/Metadata/ProcedureParameter.cs
@@ -5,8 +5,8 @@
         public string Name { get; set; }
         public string StoreType { get; set; }
         public int? Length { get; set; }
-        public byte? Precision { get; set; }
-        public byte? Scale { get; set; }
+        public int? Precision { get; set; }
+        public int? Scale { get; set; }
         public int Ordinal { get; set; }
         public bool Output { get; set; }
         public bool Nullable { get; set; }
